move decorator to jobs lib
